core/message-classification/src/main/java/uk/gov/dwp/queue/triage/core/classification/server/configuration/MessageClassificationProperties.java
package uk.gov.dwp.queue.triage.core.classification.server.configuration;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.stereotype.Component;

import java.util.concurrent.TimeUnit;

@Component
@ConfigurationProperties("messageClassification")
public class MessageClassificationProperties {

    private long initialDelay;
    private long frequency;
    private TimeUnit timeUnit;

    public long getInitialDelay() {
        return initialDelay;
    }

    public void setInitialDelay(long initialDelay) {
        this.initialDelay = initialDelay;
    }

    public long getFrequency() {
        return frequency;
    }

    public void setFrequency(long frequency) {
        this.frequency = frequency;
    }

    public TimeUnit getTimeUnit() {
        return timeUnit;
    }

    public void setTimeUnit(TimeUnit timeUnit) {
        this.timeUnit = timeUnit;
    }
}
